@@ -5,6 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using oMediaCenter.Interfaces;
 using oMediaCenter.MetaDatabase;
+using oMediaCenter.Web.Hubs;
 using oMediaCenter.Web.Model;
 //using ShowInfo;
 
@@ -22,7 +23,7 @@ namespace oMediaCenter.Web
       if (env.IsEnvironment("Development"))
       {
         // This will push telemetry data through Application Insights pipeline faster, allowing you to view results immediately.
-        builder.AddApplicationInsightsSettings(developerMode: true);
+        //        builder.AddApplicationInsightsSettings(developerMode: true);
       }
 
       builder.AddEnvironmentVariables();
@@ -38,12 +39,16 @@ namespace oMediaCenter.Web
       services.AddDbContext<MediaCenterContext>();
       services.AddDbContext<MetaDataContext>();
 
+      services.AddLogging(builder => builder.AddConsole().AddDebug());
+
       // Add framework services.
-      services.AddApplicationInsightsTelemetry(Configuration);
+      //services.AddApplicationInsightsTelemetry(Configuration);
 
       services.AddMvc();
+      services.AddSignalR();
 
       services.AddOptions();
+      services.AddSingleton<ClientConnectionDictionary>();
       services.AddSingleton<IFileReaderPluginLoader, SimpleFileReaderPluginLoader>();
 
       //services.AddTransient<IAliasProvider, AliasProvider>();
@@ -64,13 +69,18 @@ namespace oMediaCenter.Web
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline
     public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
     {
-      loggerFactory.AddConsole(Configuration.GetSection("Logging"));
-      loggerFactory.AddDebug();
+      //loggerFactory.AddConsole(Configuration.GetSection("Logging"));
+      //loggerFactory.AddDebug();
 
       app.UseDefaultFiles();
       app.UseStaticFiles();
-
-      app.UseMvc();
+      app.UseRouting();
+      app.UseEndpoints(endpoints =>
+      {
+        endpoints.MapHub<CommandHub>("/commandHub");
+        endpoints.MapControllers();
+      }
+      );
     }
   }
 }
